+++ spec/guess_spec.rb
@@ -33,6 +33,18 @@ describe 'user' do
     expect(@b.driver.find_elements(locatine: 'collection span').length).to eq 3
   end
 
+  it 'fails to find anyting if there is nothing' do
+    @b.goto page(16)
+    expect{@b.driver.find_element(locatine: 'never span')}.to raise_error Selenium::WebDriver::Error::NoSuchElementError
+    expect(@b.driver.find_elements(locatine: 'nothing label').length).to eq 0
+  end
+
+  it 'fails to find anyting if there is something strange' do
+    @b.goto page(16)
+    expect{@b.driver.find_element(locatine: 'strange one')}.to raise_error Selenium::WebDriver::Error::NoSuchElementError
+    expect(@b.driver.find_elements(locatine: 'not good').length).to eq 0
+  end
+
   after(:all) do
     FileUtils.remove_dir('./locatine_files/', true)
     make_request('http://localhost:7733/locatine/stop')